+++ lib/buttercut.rb
@@ -1,6 +1,7 @@
 require_relative 'buttercut/fcpx'
 require_relative 'buttercut/fcp7'
 require_relative 'buttercut/recipe'
+require_relative 'buttercut/broll_manifest'
 
 # ButterCut - Video editor XML generator
 #